Updated webBoard README file
WebBoardPoC:
This module is a proof of concept for WebBoard.
You can see the running demo here on the heruko cloud:
http://morning-taiga-2630.herokuapp.com/

To run this application on the local machine, do following changes:
$ cd webRTCPoC
$ sudo npm install express
$ vim index.js
$ nodejs index.js (node index.js on fedora)

This demo demonstrate the single user and single client scribbling application. Since the
server uses single buffer right now, hence only one user can see the actual stokes.

The design will be changes the to support multiple users and multiple images with live and
recorded storage.

Signed-off-by: Ashwin Vasani <akvasani@asu.edu>